+++ src/main/java/com/example/habittracker/Service/HabitService.java
@@ -17,18 +17,28 @@ public class HabitService {
     private final UserHabitRepository userHabitRepository;
     private final HabitHistoryRepository habitHistoryRepository;
     private final ChallengeRepository challengeRepository;
+    private final UserChallengeRepository userChallengeRepository;
+    private final ChallengeProgressService challengeProgressService;
 
-    public HabitService(HabitRepository habitRepository, UserService userService, UserHabitRepository userHabitRepository, HabitHistoryRepository habitHistoryRepository, ChallengeRepository challengeRepository) {
+    public HabitService(HabitRepository habitRepository, UserService userService, UserHabitRepository userHabitRepository, HabitHistoryRepository habitHistoryRepository, ChallengeRepository challengeRepository, UserChallengeRepository userChallengeRepository, ChallengeProgressService challengeProgressService) {
         this.habitRepository = habitRepository;
         this.userService = userService;
         this.userHabitRepository = userHabitRepository;
         this.habitHistoryRepository = habitHistoryRepository;
         this.challengeRepository = challengeRepository;
+        this.userChallengeRepository = userChallengeRepository;
+        this.challengeProgressService = challengeProgressService;
     }
 
+
+//    public List<UserHabit> getUserHabits(User user) {
+//        return this.userHabitRepository.findHabitsForUser(user.getUserId()).get();
+//    }
     @Transactional
     public List<UserHabit> getUserHabits(User user) {
-        return this.userHabitRepository.findHabitsForUser(user.getUserId()).get();
+        return this.userHabitRepository.findHabitsForUser(user.getUserId()).stream()
+                .filter(userHabit -> !userHabit.isUnavailable())
+                .collect(Collectors.toList());
     }
 
     @Transactional
@@ -64,7 +74,24 @@ public class HabitService {
         userHabit.setTargetCount(createHabit.getTargetCount());
         userHabit.setCurrentCount(0L);
         userHabit.setDifficulty(createHabit.getDifficulty());
+        if(habitDTO.getType() == Habit.HabitType.NEGATIVE){
+            userHabit.setCompleted(true);
+        }else{
+            userHabit.setCompleted(false);
+        }
         this.userHabitRepository.save(userHabit);
+
+        if (habitDTO.getType() == Habit.HabitType.NEGATIVE) {
+            HabitHistory initialHistory = HabitHistory.builder()
+                    .userHabit(userHabit)
+                    .date(LocalDate.now())
+                    .isCompleted(userHabit.isCompleted())
+                    .positiveCount(0L)
+                    .negativeCount(0L)
+                    .expEarned(0L)
+                    .build();
+            this.habitHistoryRepository.save(initialHistory);
+        };
     }
 
     @Transactional
@@ -98,13 +125,39 @@ public class HabitService {
 
         habit.setTitle(habitDTO.getTitle());
         habit.setDescription(habitDTO.getDescription());
-        habit.setType(habitDTO.getType());
         habit.setChallenge(challenge);
-        this.habitRepository.save(habit);
+
         UserHabit userHabit = this.userHabitRepository.findUserHabitByHabitAndUser(habit,user).orElseThrow(()->new RuntimeException("Lỗi khi lưu dữ liệu chỉnh sửa!"));
+        Habit.HabitType oldType = habit.getType();
+        Habit.HabitType newType = habitDTO.getType();
+
+        if(oldType != newType){
+            if(newType == Habit.HabitType.NEGATIVE){
+                userHabit.setCompleted(!(oldType == Habit.HabitType.BOTH && userHabit.getNegativeCount() >= userHabit.getTargetCount()));
+                userHabit.setPositiveCount(0L);
+                HabitHistory initialNegativeHistory = this.habitHistoryRepository.findDailyHistory(userHabit, LocalDate.now())
+                        .orElseGet(() -> HabitHistory.builder()
+                                .userHabit(userHabit)
+                                .date(LocalDate.now())
+                                .build());
+                initialNegativeHistory.setCompleted(userHabit.isCompleted());
+                initialNegativeHistory.setPositiveCount(0L);
+                initialNegativeHistory.setNegativeCount(0L);
+                initialNegativeHistory.setExpEarned(0L);
+                this.habitHistoryRepository.save(initialNegativeHistory);
+            } else if (newType == Habit.HabitType.POSITIVE) {
+                userHabit.setNegativeCount(0L);
+                userHabit.setCompleted(oldType == Habit.HabitType.BOTH && userHabit.getPositiveCount() >= userHabit.getTargetCount());
+            } else if (newType == Habit.HabitType.BOTH) {
+                userHabit.setCompleted(oldType == Habit.HabitType.POSITIVE && userHabit.getPositiveCount() - userHabit.getNegativeCount() >= userHabit.getTargetCount());
+            }
+        }
         userHabit.setTargetCount(habitDTO.getTargetCount());
         userHabit.setDifficulty(habitDTO.getDifficulty());
         this.userHabitRepository.save(userHabit);
+
+        habit.setType(habitDTO.getType());
+        this.habitRepository.save(habit);
     }
 
     @Transactional
@@ -113,9 +166,23 @@ public class HabitService {
         Habit habit = getHabit(habitId);
         UserHabit userHabit = this.userHabitRepository.findUserHabitByHabitAndUser(habit,user).orElseThrow(()->new RuntimeException("Lỗi xảy ra khi xóa thói quen!"));
         List<HabitHistory> habitHistories = this.habitHistoryRepository.findAllByUserHabit(userHabit);
-        this.habitHistoryRepository.deleteAll(habitHistories);
-        this.userHabitRepository.delete(userHabit);
-        this.habitRepository.delete(habit);
+
+        if(habit.getChallenge() != null){
+            UserChallenge userChallenge = this.userChallengeRepository.findByUserAndChallenge(user,habit.getChallenge()).orElseThrow(()->new RuntimeException("Không tìm thấy dữ liệu để xóa!"));
+            if(userChallenge.getStatus() == UserChallenge.Status.COMPLETE){
+                userHabit.setUnavailable(true);
+                this.userHabitRepository.save(userHabit);
+            }else{
+                this.habitHistoryRepository.deleteAll(habitHistories);
+                this.userHabitRepository.delete(userHabit);
+                this.habitRepository.delete(habit);
+            }
+        }else{
+            this.habitHistoryRepository.deleteAll(habitHistories);
+            this.userHabitRepository.delete(userHabit);
+            this.habitRepository.delete(habit);
+        }
+
     }
 
     @Transactional
@@ -156,7 +223,7 @@ public class HabitService {
         habitDTO.setNegativeCount(userHabit.getNegativeCount());
 
         if((habit.getType().name().equals("BOTH") && userHabit.getTargetCount() <= userHabit.getPositiveCount() - userHabit.getNegativeCount())
-        || (habit.getType().name().equals("NEGATIVE") && userHabit.getTargetCount() <= userHabit.getNegativeCount())
+        || (habit.getType().name().equals("NEGATIVE") && userHabit.getTargetCount() > userHabit.getNegativeCount())
         ||(habit.getType().name().equals("POSITIVE") && userHabit.getTargetCount() <= userHabit.getPositiveCount())){
             userHabit.setCompleted(true);
             habitHistory.setCompleted(true);
@@ -170,6 +237,14 @@ public class HabitService {
         this.userHabitRepository.save(userHabit);
         this.habitHistoryRepository.save(habitHistory);
 
+        if(habit.getChallenge() != null) {
+            UserChallenge userChallenge = this.userChallengeRepository.findByUserAndChallenge(user, habit.getChallenge()).orElse(null);
+            if(userChallenge != null && userChallenge.getStatus() == UserChallenge.Status.ACTIVE){
+                this.challengeProgressService.calculateAndSaveDailyProgress(userChallenge.getUserChallengeId(),today);
+                this.challengeProgressService.recalculateUserChallengeProgress(userChallenge);
+                this.challengeProgressService.updateChallengeStreak(userChallenge);
+            }
+        }
 
         return habitDTO;
     }
@@ -196,8 +271,26 @@ public class HabitService {
 
             userHabit.setNegativeCount(0L);
             userHabit.setPositiveCount(0L);
-            userHabit.setCompleted(false);
+            userHabit.setCompleted(userHabit.getHabit().getType() == Habit.HabitType.NEGATIVE);
             this.userHabitRepository.save(userHabit);
+        }
+    }
+
+    @Transactional
+    public void setHabitHistoryNewDay(){
+        List<UserHabit> userHabits = this.userHabitRepository.findAll();
+        LocalDate today = LocalDate.now();
+        for(UserHabit userHabit : userHabits){
+            HabitHistory habitHistory = this.habitHistoryRepository.findByUserHabitAndDate(userHabit,today)
+                    .orElseGet(()->new HabitHistory().builder()
+                            .userHabit(userHabit)
+                            .negativeCount(userHabit.getNegativeCount())
+                            .positiveCount(userHabit.getPositiveCount())
+                            .date(today)
+                            .isCompleted(userHabit.isCompleted())
+                            .build()
+                    );
+            habitHistoryRepository.save(habitHistory);
         }
     }
